chore: generators refactor
