+++ src/generation/index.ts
@@ -1,7 +1,7 @@
 import { TypeScriptTypettaPluginConfig } from './config'
-import { TsTypettaGenerator } from './dao-generator'
-import { InputTypettaGenerator } from './input-generator'
-import { ResolverTypettaGenerator } from './resolver-generator'
+import { TsTypettaGenerator } from './generators/dao-generator'
+import { InputTypettaGenerator } from './generators/input-generator'
+import { ResolverTypettaGenerator } from './generators/resolver-generator'
 import { TsTypettaGeneratorNode, TsTypettaGeneratorScalar } from './types'
 import { TsTypettaVisitor } from './visitor'
 import { Types, PluginFunction, PluginValidateFn } from '@graphql-codegen/plugin-helpers'
