+++ tests/mongodb/dao.mock.ts
@@ -2,10 +2,6 @@ import * as T from '../../src'
 import * as types from './models.mock'
 import * as M from 'mongodb'
 
-//--------------------------------------------------------------------------------
-//----------------------------------- ADDRESS ------------------------------------
-//--------------------------------------------------------------------------------
-
 export type AddressExcludedFields = never
 export type AddressRelationFields = 'cities'
 
@@ -109,10 +105,6 @@ export class InMemoryAddressDAO<MetadataType, OperationMetadataType> extends T.A
   }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------ AUDIT -------------------------------------
-//--------------------------------------------------------------------------------
 
 export type AuditExcludedFields = never
 export type AuditRelationFields = never
@@ -223,10 +215,6 @@ export class InMemoryAuditDAO<MetadataType, OperationMetadataType> extends T.Abs
 
 
 
-//--------------------------------------------------------------------------------
-//---------------------------------- AUDITABLE -----------------------------------
-//--------------------------------------------------------------------------------
-
 export function auditableSchema(): T.Schema<types.Scalars> {
   return {
     'createdBy': {
@@ -277,10 +265,6 @@ export type AuditableInsert = {
 }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------- CITY -------------------------------------
-//--------------------------------------------------------------------------------
 
 export type CityExcludedFields = 'computedAddressName' | 'computedName'
 export type CityRelationFields = never
@@ -393,10 +377,6 @@ export class InMemoryCityDAO<MetadataType, OperationMetadataType> extends T.Abst
   }
 
 
-
-//--------------------------------------------------------------------------------
-//----------------------------- DEFAULTFIELDSENTITY ------------------------------
-//--------------------------------------------------------------------------------
 
 export type DefaultFieldsEntityExcludedFields = never
 export type DefaultFieldsEntityRelationFields = never
@@ -534,10 +514,6 @@ export class InMemoryDefaultFieldsEntityDAO<MetadataType, OperationMetadataType>
 
 
 
-//--------------------------------------------------------------------------------
-//------------------------------------ DEVICE ------------------------------------
-//--------------------------------------------------------------------------------
-
 export type DeviceExcludedFields = never
 export type DeviceRelationFields = 'user'
 
@@ -648,10 +624,6 @@ export class InMemoryDeviceDAO<MetadataType, OperationMetadataType> extends T.Ab
   }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------- DOG --------------------------------------
-//--------------------------------------------------------------------------------
 
 export type DogExcludedFields = never
 export type DogRelationFields = 'owner'
@@ -765,10 +737,6 @@ export class InMemoryDogDAO<MetadataType, OperationMetadataType> extends T.Abstr
 
 
 
-//--------------------------------------------------------------------------------
-//--------------------------------- EMBEDDEDUSER ---------------------------------
-//--------------------------------------------------------------------------------
-
 export function embeddedUserSchema(): T.Schema<types.Scalars> {
   return {
     'e': { embedded: () => embeddedUser2Schema(), array: true },
@@ -797,10 +765,6 @@ export type EmbeddedUserInsert = {
 
 
 
-//--------------------------------------------------------------------------------
-//-------------------------------- EMBEDDEDUSER2 ---------------------------------
-//--------------------------------------------------------------------------------
-
 export function embeddedUser2Schema(): T.Schema<types.Scalars> {
   return {
     'user': { relation: () => userSchema(), required: true },
@@ -823,10 +787,6 @@ export type EmbeddedUser2Insert = {
 
 
 
-//--------------------------------------------------------------------------------
-//-------------------------------- EMBEDDEDUSER3 ---------------------------------
-//--------------------------------------------------------------------------------
-
 export function embeddedUser3Schema(): T.Schema<types.Scalars> {
   return {
     'user': { relation: () => userSchema() },
@@ -847,10 +807,6 @@ export type EmbeddedUser3Insert = {
 }
 
 
-
-//--------------------------------------------------------------------------------
-//-------------------------------- EMBEDDEDUSER4 ---------------------------------
-//--------------------------------------------------------------------------------
 
 export function embeddedUser4Schema(): T.Schema<types.Scalars> {
   return {
@@ -873,10 +829,6 @@ export type EmbeddedUser4Insert = {
 
 
 
-//--------------------------------------------------------------------------------
-//-------------------------------- EMBEDDEDUSER5 ---------------------------------
-//--------------------------------------------------------------------------------
-
 export function embeddedUser5Schema(): T.Schema<types.Scalars> {
   return {
     'userId': {
@@ -895,10 +847,6 @@ export type EmbeddedUser5Insert = {
 }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------ HOTEL -------------------------------------
-//--------------------------------------------------------------------------------
 
 export type HotelExcludedFields = never
 export type HotelRelationFields = never
@@ -1108,10 +1056,6 @@ export class InMemoryHotelDAO<MetadataType, OperationMetadataType> extends T.Abs
 
 
 
-//--------------------------------------------------------------------------------
-//--------------------------------- MOCKEDENTITY ---------------------------------
-//--------------------------------------------------------------------------------
-
 export type MockedEntityExcludedFields = never
 export type MockedEntityRelationFields = 'user'
 
@@ -1223,10 +1167,6 @@ export class InMemoryMockedEntityDAO<MetadataType, OperationMetadataType> extend
   }
 
 
-
-//--------------------------------------------------------------------------------
-//--------------------------------- ORGANIZATION ---------------------------------
-//--------------------------------------------------------------------------------
 
 export type OrganizationExcludedFields = 'computedName'
 export type OrganizationRelationFields = never
@@ -1346,10 +1286,6 @@ export class InMemoryOrganizationDAO<MetadataType, OperationMetadataType> extend
   }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------- POST -------------------------------------
-//--------------------------------------------------------------------------------
 
 export type PostExcludedFields = never
 export type PostRelationFields = 'author'
@@ -1497,10 +1433,6 @@ export class InMemoryPostDAO<MetadataType, OperationMetadataType> extends T.Abst
 
 
 
-//--------------------------------------------------------------------------------
-//--------------------------------- POSTMETADATA ---------------------------------
-//--------------------------------------------------------------------------------
-
 export function postMetadataSchema(): T.Schema<types.Scalars> {
   return {
     'region': {
@@ -1526,10 +1458,6 @@ export type PostMetadataInsert = {
 }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------- USER -------------------------------------
-//--------------------------------------------------------------------------------
 
 export type UserExcludedFields = never
 export type UserRelationFields = 'dogs' | 'friends'
@@ -1777,10 +1705,6 @@ export class InMemoryUserDAO<MetadataType, OperationMetadataType> extends T.Abst
 
 
 
-//--------------------------------------------------------------------------------
-//-------------------------------- USERCOLLECTION --------------------------------
-//--------------------------------------------------------------------------------
-
 export function userCollectionSchema(): T.Schema<types.Scalars> {
   return {
     'users': { relation: () => userSchema(), required: true, array: true },
@@ -1803,10 +1727,6 @@ export type UserCollectionInsert = {
 }
 
 
-
-//--------------------------------------------------------------------------------
-//------------------------- USERNAMEPASSWORDCREDENTIALS --------------------------
-//--------------------------------------------------------------------------------
 
 export function usernamePasswordCredentialsSchema(): T.Schema<types.Scalars> {
   return {
@@ -1990,10 +1910,6 @@ export class EntityManager<MetadataType = never, OperationMetadataType = never, 
 
 }
 
-
-//--------------------------------------------------------------------------------
-//------------------------------------- UTILS ------------------------------------
-//--------------------------------------------------------------------------------
 
 type DAOName = keyof DAOGenericsMap<never, never>
 type DAOGenericsMap<MetadataType, OperationMetadataType> = {
